Extract Keystore to a file
